+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Conductor: Magento 2 Platform Support Changelog
 ==============================================
 
+# 0.9.1
+- Updated conductor/application-orchestration require to ~0.9.3
+
+# 0.9.0
+- Tagging for initial consistency with other modules
+
 # 0.2.0 (Unreleased)
 - Renamed to Conductor
 - Updated PHP version requirement to 7.1